+++ meow/include/meow/repository/repository.hpp
@@ -26,6 +26,7 @@ struct RepositoryPackage {
     std::vector<types::PackageName> provides;
     std::vector<types::PackageName> conflicts;
     std::vector<types::PackageName> depends;
+    std::vector<types::OptionalDependency> optionalDepends;
     std::vector<RepositoryVersion> versions;
 };